Implementing jwt tokens for login

+++ src/main/java/main/repository/StoryInfoRepository.java
@@ -1,12 +1,10 @@
 package main.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import main.pojo.Story;
 
-public interface StoryInfoRepository extends JpaRepository<Story, Integer> {
-//
-//	public void findByAssignedUserId();
-//
-//	public void findByReportingUserId();
+public interface StoryInfoRepository extends JpaRepository<Story, Long> {
+	
 
 }
